// src/api/generated/types.gen.ts
export type Problem = {
    type?: string;
    title?: string;
    status?: number;
    detail: string;
    code?: string;
    instance?: string;
    errors?: Array<ValidationError>;
};

export type ValidationError = {
    field: string;
    message: string;
};